Client subscriptions are now implemented.
Updated loadtest tool to test pub and sub, and confirm that all messages
get through.

Added host argument to loadtest tool, so it can test other servers.
Using this, confirmed that this client can talk to mosquitto.

Improved logging and ignored common useless errors.

Fixed a bug with tincorrect formatting of Subscribe messages. Fixed
our server to die on the same thing as mosquitto was.

Fixed a bug causing data already submitted to the kernel to be lost
if the connection was closed from the writer thread (is this a bug
in Go? in Linux?). Now we rely on the remote server to close the connection
to cause the client reader goroutine to exit. Timeouts will improve this
situation.

Exported errors for each type of connection error, and use them various
places.